Update README.md
The purpose of this project was to build a conversion funnel with cohorts 
in order to calculate the retention rate of the store. 
Our conversion funnel shows how many customers viewed products online, added
items to their cart, and purchased items from the online store. As each month progresses, 
you will see that our retention rate drops. There may be a possibility that the store doesn't offer memberships.